Problem: Virtual Instruction Set Architecture required for Intermediary Representation
Solution: introduce the bare essentials of a clear text parsing of the visa. The lingua franca of sio is sio_visa. Other people may create
frontends that target sio_visa if they should so wish.

Try to keep the primitives as close to rust's primitives as possible. This is heavily inspired by webassembly but doesn't do stupid things
like conflating the concept of a module and a process. The IR will expose more language concepts such as concurrency out of the box.